move collect_shared_memory_operands
